requirements.txt
python-telegram-bot==20.7
PyGithub==2.1.1
python-dotenv==1.0.0
aiohttp==3.9.1
requests==2.31.0
SQLAlchemy==2.0.23
aiosqlite==3.15.0
pydantic==2.5.0